Add UV Dependabot Configuration (#22)
# Pull Request

## Description

This pull request includes changes to the `.github/dependabot.yml` file
to update the configuration for dependency updates. The most important
changes include the removal of the commit message prefix for GitHub
Actions updates and the addition of a new package ecosystem
configuration for "uv".

Changes to `.github/dependabot.yml`:

* Removed the `commit-message` prefix for GitHub Actions updates.
* Added a new package ecosystem configuration for "uv" with a daily
update schedule, targeting the "main" branch, and specifying the
timezone as "Europe/London".

Fixes #15