Get Pull Requests (#34)
# Pull Request

## Description

This pull request includes significant changes to the
`DependabotTrigger` project, focusing on updating dependencies,
refactoring the main script, and modifying the build configuration. The
most important changes are outlined below:

### Dependency Management:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L5-R7):
Added `pygithub` as a dependency to the project.

### Code Refactoring:

*
[`src/__main__.py`](diffhunk://#diff-5745cb7d8ed87db31cdaef7ff6d897c0bcaf2200967146f11d0e1728540ea0b5R1-R4):
Introduced the `app` function from `src/app.py` and set it as the entry
point for the script.
*
[`src/app.py`](diffhunk://#diff-04791d82dd15fdd480f084d7ef65a10789fa5012cb7935f76080763444d48a00R1-R25):
Created a new `app` function that prints a greeting and retrieves open
pull requests from a specified repository using the `pygithub` library.
*
[`src/main.py`](diffhunk://#diff-2e5ad92c43aa96cc3a9cef6c6aec998b216f1379c43b1f651013d25e55989312L1-L6):
Removed the old `main` function and its invocation.

### Build Configuration:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL1-R7):
Updated the `run` command to use `uv run python -m src` instead of a
script name variable.

Fixes #29