Log Pull Request Links (#59)
# Pull Request

## Description

This pull request introduces enhancements to the logging and
functionality of the `app.py` script. The changes improve the clarity of
the code, add a new logging function, and adjust the logging level for
pull request details.

### Functional Enhancements:
* Added a new function, `log_all_pull_requests`, which logs the
repository name, total count of pull requests, and their URLs. This
function is called in `app()` to provide more comprehensive logging of
pull request details.
[[1]](diffhunk://#diff-04791d82dd15fdd480f084d7ef65a10789fa5012cb7935f76080763444d48a00L17-R20)
[[2]](diffhunk://#diff-04791d82dd15fdd480f084d7ef65a10789fa5012cb7935f76080763444d48a00R51-R61)

### Logging Improvements:
* Changed the logging level for individual pull request details from
`info` to `debug` in `get_pull_requests`, reducing noise in the logs for
non-critical information.