Add CodeLimit Job (#65)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow to
integrate CodeLimit, a tool for enforcing code limits.

Key change:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R183-R197):
Introduced a new `run-codelimit` job that checks code limits using the
`getcodelimit/codelimit-action`. This job includes steps for checking
out the repository and running CodeLimit, with appropriate permissions
for contents and pull requests.