Add Project Description (#78)
# Pull Request

## Description

This pull request updates the `README.md` file to include a brief
description of the `DependabotTrigger` script. The description explains
its purpose and functionality, providing context for users.

Documentation update:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R3-R4):
Added a description of the `DependabotTrigger` script, explaining that
it is a Python tool that uses Playwright to automate triggering
Dependabot version update jobs for GitHub repositories.