Update Dependabot Groups and Configuration (#80)
# Pull Request

## Description

This pull request updates the Dependabot configuration to enhance
scheduling flexibility and refine group-specific rules. The key changes
include adding a timezone for the cron schedule and introducing new
attributes for update groups.

### Scheduling enhancements:
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R11-R19):
Added a `timezone` field with the value "Europe/London" to the schedule
configuration, ensuring the cron job executes in the specified timezone.
[[1]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R11-R19)
[[2]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R30-R34)

### Group-specific refinements:
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R11-R19):
Added the `applies-to` attribute with the value "version-updates" for
`github-actions` and `uv` groups to specify the type of updates these
groups apply to.
[[1]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R11-R19)
[[2]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R30-R34)
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R11-R19):
Introduced `exclude-patterns` for the `github-actions` group to exclude
updates matching the pattern "super-linter/*".