Remove Pinact Configuration (#85)
# Pull Request

## Description

This pull request removes the use of a custom configuration file for
Pinact and updates related commands to use the default configuration.
The most important changes are as follows:

### Removal of custom Pinact configuration:
* Deleted the `.github/other-configurations/pinact.yml` file, which
previously defined a custom Pinact configuration, including version and
ignored actions.

### Updates to Pinact commands:
* Modified the `pinact-run`, `pinact-check`, and `pinact-update`
commands in the `Justfile` to no longer reference the deleted custom
configuration file, defaulting to the standard Pinact configuration
instead.