Fix Zizmor SARIF Command (#88)
# Pull Request

## Description

This pull request updates the workflow and Justfile to streamline the
setup of dependencies and improve the execution of the `zizmor` tool.
The most important changes include replacing the custom Python
dependency setup with specific tools (`uv` and `Just`), and introducing
a new command for generating SARIF output.

### Workflow updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L89-R94):
Replaced the custom Python dependency setup action with
`astral-sh/setup-uv` and added `extractions/setup-just` for setting up
`uv` and `Just`, respectively. Updated the `Run zizmor` step to use the
new `just zizmor-check-sarif` command.

### Justfile updates:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL98-R102):
Modified the `zizmor-check` command to use `uvx` instead of `zizmor`.
Added a new `zizmor-check-sarif` command to generate SARIF output for
`zizmor` checks.