Sort using case-insensitive repository names (#90)
# Pull Request

## Description

This pull request includes a version update for a development dependency
and a minor improvement to a sorting function in the application code.

Dependency update:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L14-R14):
Updated the `ruff` dependency from version `0.11.8` to `0.11.9` in the
`dev` optional dependencies.

Code improvement:

*
[`src/app.py`](diffhunk://#diff-04791d82dd15fdd480f084d7ef65a10789fa5012cb7935f76080763444d48a00L69-R69):
Modified the sorting function in `get_all_repos` to sort repository
names in a case-insensitive manner by using `.lower()` on
`repo.full_name`.