Bump SuperLinter to 7.4.0 (#93)
# Pull Request

## Description

This pull request updates the GitHub Actions workflow configuration to
improve code linting by upgrading the `super-linter` version and
simplifying the environment variables.

Changes to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L28-L35):
Updated `super-linter` to version 7.4.0 by changing the `uses` reference
to a newer commit hash. This ensures the workflow uses the latest
linting features and fixes.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L28-L35):
Removed the `EDITORCONFIG_FILE_NAME` environment variable, as it is no
longer required for the linting process.