Migrate Sync Labels to Common Sync Labels (#100)
# Pull Request

## Description

This pull request updates the GitHub Actions workflow for syncing labels
to use a reusable workflow, simplifies permissions, and introduces a
manual trigger. The most important changes are as follows:

### Workflow improvements:
* Added `workflow_dispatch` to allow manual triggering of the workflow.
* Replaced inline job steps with a reusable workflow from
`JackPlowman/reusable-workflows`. This simplifies the configuration and
centralizes the logic for syncing labels.

### Permissions updates:
* Changed `permissions` at the workflow level to `{}` (no default
permissions) and specified granular permissions (`contents: read` and
`pull-requests: write`) at the job level.