Update SuperLinter Configuration (#161)
# Pull Request

## Description

This pull request updates the `.github/workflows/code-checks.yml` file
to adjust the configuration of the code checks workflow. The most
significant change is the addition of new validation options for
`editorconfig` and `git commitlint`, while removing the validation for
`pyink`.

### Updates to code checks workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R38-L44):
Added `VALIDATE_EDITORCONFIG` and `VALIDATE_GIT_COMMITLINT`
configuration options, both set to `false`, to expand the scope of tools
included in the workflow.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R38-L44):
Removed the `VALIDATE_PYTHON_PYINK` configuration option, indicating
that `pyink` validation is no longer part of the workflow.