Add actionlint Git Hook (#165)
# Pull Request

## Description

This pull request introduces a new `actionlint-check` task to the
codebase and integrates it into the pre-commit hooks. The changes ensure
that GitHub Actions workflows are linted as part of the development
process.

### Additions to GitHub Actions linting:

* **New `actionlint-check` task in `Justfile`:** Added a task to run
`actionlint` for validating GitHub Actions workflows.
* **Integration into `lefthook.yml`:** Included `Actionlint Checks` in
the pre-commit hooks to automatically run the `actionlint-check` task
during pre-commit validation.