Update Reusable Workflows (#170)
# Pull Request

## Description

This pull request updates several GitHub Actions workflows to use a
newer version of reusable workflows. The changes ensure that the
workflows reference the latest version (`v2025.08.04.01`) of the
reusable workflows from `JackPlowman/reusable-workflows`.

### Workflow updates:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL15-R15):
Updated the reusable workflow reference to `common-clean-caches.yml` to
the latest version.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L54-R54):
Updated the reusable workflow references for both
`common-code-checks.yml` and `codeql-analysis.yml` to the latest
version.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L54-R54)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L66-R66)
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL15-R15):
Updated the reusable workflow reference to
`common-pull-request-tasks.yml` to the latest version.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L19-R19):
Updated the reusable workflow reference to `common-sync-labels.yml` to
the latest version.